% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC.evaluation.R
\name{print.pc_measures}
\alias{print.pc_measures}
\title{print function for PC.evaluation results}
\usage{
\method{print}{pc_measures}(x, round = 3, ...)
}
\description{
print function for PC.evaluation results
}
