% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PC.evaluation.R
\name{PC.evaluation}
\alias{PC.evaluation}
\title{Evaluate a partly conditional model}
\usage{
PC.evaluation(pc.object, newdata, prediction.time = NULL,
  conditioning.time.window = NULL, risk.threshold, pnf.threshold = 0.5,
  pcf.threshold = 0.25, bootstraps = 500, alpha = 0.05, silent = FALSE)
}
\arguments{
\item{pc.object}{output from PC.Cox or PC.GLM functions in the 'partlyconditional' R package used to fit partly conditional Cox/GLM models.}

\item{newdata}{data.frame with new data for which to estimate summary measures. All variables used to fit the PC.Cox/PC.GLM model must be present. Observations with missing data will be removed.}

\item{prediction.time}{numeric value of prediction time (from conditioning.time) to estimate future risk. Prediction time should be on the same scale as the measurement time and the survival times provided to fit the partly conditional model.}

\item{conditioning.time.window}{Time. All measurement times in newdata exceeding this conditioning time will be removed from analysis (and a message will be produced if silent = FALSE).}

\item{risk.threshold}{numeric threshold on the risk scale used to classify individuals as 'high-risk' for TPF, FPF, NPV, and PPV measures.}

\item{pnf.threshold}{threshold q to estimate the proportion needed to follow PNF(q). Defaults to q = .5. PNF(q), is the proportion of the population at highest risk that one needs classify high risk in order that a proportion q of the cases will be identified.}

\item{pcf.threshold}{threshold p to estimate the proportion of cases followed PCF(p) measure. Defaults to p = .25. PCF(p) is the proportion of cases included in the proportion p of individuals in the population at highest risk.}

\item{bootstraps}{Number of bootstraps used for confidence intervals and standard error estimation. Default is 500. Set to 0 if no CI's are desired. See details below for further information on CI construction.}

\item{alpha}{alpha level for confidence interval calculations. Default is 0.05 for 95\% confidence intervals.}

\item{silent}{set to TRUE to hide messages printed from function. Default is silent = FALSE.}
}
\value{
An object of class 'pc_evaluate' which is a list containing:

\item{measures}{tibble consisting of estimates for prediction error, AUC, true positive fraction (TPF), false positive fraction (FPF), positive predictive value (PPV), negative predictive value (NPV), proportion of cases followed (PCF), proportion needed to follow-up (PNF), proportion high risk, and outcome prevalence.  }
\item{roc}{ tibble with components of the roc curve including, TPF, FPF, PPV, NPV, risk.threshold, and risk.percentile, at all risk thresholds observed.}
\item{data.for.measures}{data used to , consists of all observations with measurement time within the conditioning.time.window window. }
\item{call, bootstrap.info}{Inputs from function call. }
}
\description{
Evaluate the predictive performance of a partly conditional (PC) model fit using the 'partlyconditional' R package.  Given validation data consisting of a time-to-event outcome and longitudinal marker information on a set of individuals up to a fixed time window (conditioning.time.window), this function estimates measures of prognostic accuracy for a specified future landmark prediction time ('prediction.time'). Bootstrap confidence intervals are provided for measures at the risk threshold provided.
}
\details{

}
\examples{

library(partlyconditional)
library(longsurvAccuracyMeasures)
data(pc_data)

pc.model.1 <-  PC.Cox(
 id = "sub.id",
 stime = "time",
 status = "status",
 measurement.time = "meas.time",
 markers = c("marker", "marker_2"),
 data = pc_data,
 use.BLUP = c(FALSE, FALSE),
 knots.measurement.time = NA)

pc.model.1

result <- PC.evaluation(pc.cox.1,
                       newdata = pc_data,
                       conditioning.time.window = c(18, 24), #make predictions using data up to time 18-24
                       prediction.time = 12,  #one year predictions, conditional on historical data
                       risk.threshold = .4,
                       pnf.threshold = .5,
                       pcf.threshold = .18,
                       bootstraps = 25) #should use more bootstrap replicates in practice.

result

}
